# graph: replace custom reverse logic with slices.Reverse