--- a/graph/community/louvain_directed_test.go
+++ b/graph/community/louvain_directed_test.go
@@ -7,6 +7,7 @@ package community
 import (
 	"math"
 	"reflect"
+	"slices"
 	"sort"
 	"testing"
 
@@ -631,7 +632,7 @@ func testModularizeDirected(t *testing.T, test communityDirectedQTest, g graph.D
 		}
 
 		// Recovery of Q values is reversed.
-		if reverse(qs); !sort.Float64sAreSorted(qs) {
+		if slices.Reverse(qs); !sort.Float64sAreSorted(qs) {
 			t.Errorf("Q values not monotonically increasing: %.5v", qs)
 		}
 	}