--- a/graph/community/louvain_undirected_test.go
+++ b/graph/community/louvain_undirected_test.go
@@ -7,6 +7,7 @@ package community
 import (
 	"math"
 	"reflect"
+	"slices"
 	"sort"
 	"testing"
 
@@ -694,7 +695,7 @@ func testModularizeUndirected(t *testing.T, test communityUndirectedQTest, g gra
 		}
 
 		// Recovery of Q values is reversed.
-		if reverse(qs); !sort.Float64sAreSorted(qs) {
+		if slices.Reverse(qs); !sort.Float64sAreSorted(qs) {
 			t.Errorf("Q values not monotonically increasing: %.5v", qs)
 		}
 	}