--- a/graph/topo/bron_kerbosch.go
+++ b/graph/topo/bron_kerbosch.go
@@ -5,8 +5,9 @@
 package topo
 
 import (
+	"slices"
+
 	"gonum.org/v1/gonum/graph"
-	"gonum.org/v1/gonum/graph/internal/ordered"
 	"gonum.org/v1/gonum/graph/internal/set"
 )
 
@@ -15,7 +16,7 @@ import (
 func DegeneracyOrdering(g graph.Undirected) (order []graph.Node, cores [][]graph.Node) {
 	order, offsets := degeneracyOrdering(g)
 
-	ordered.Reverse(order)
+	slices.Reverse(order)
 	cores = make([][]graph.Node, len(offsets))
 	offset := len(order)
 	for i, n := range offsets {
@@ -145,7 +146,7 @@ func BronKerbosch(g graph.Undirected) [][]graph.Node {
 	x := set.NewNodes()
 	var bk bronKerbosch
 	order, _ := degeneracyOrdering(g)
-	ordered.Reverse(order)
+	slices.Reverse(order)
 	for _, v := range order {
 		neighbours := graph.NodesOf(g.From(v.ID()))
 		nv := set.NewNodesSize(len(neighbours))