--- a/graph/topo/tarjan.go
+++ b/graph/topo/tarjan.go
@@ -6,6 +6,7 @@ package topo
 
 import (
 	"fmt"
+	"slices"
 
 	"gonum.org/v1/gonum/graph"
 	"gonum.org/v1/gonum/graph/internal/ordered"
@@ -70,12 +71,10 @@ func sortedFrom(sccs [][]graph.Node, order func([]graph.Node)) ([]graph.Node, er
 	}
 	var err error
 	if sc != nil {
-		for i, j := 0, len(sc)-1; i < j; i, j = i+1, j-1 {
-			sc[i], sc[j] = sc[j], sc[i]
-		}
+		slices.Reverse(sc)
 		err = sc
 	}
-	ordered.Reverse(sorted)
+	slices.Reverse(sorted)
 	return sorted, err
 }
 
@@ -100,12 +99,12 @@ func tarjanSCCstabilized(g graph.Directed, order func([]graph.Node)) [][]graph.N
 		}
 	} else {
 		order(nodes)
-		ordered.Reverse(nodes)
+		slices.Reverse(nodes)
 
 		succ = func(id int64) []graph.Node {
 			to := graph.NodesOf(g.From(id))
 			order(to)
-			ordered.Reverse(to)
+			slices.Reverse(to)
 			return to
 		}
 	}