@jbduncan 2024-03-23
I've made a couple new benchmarks, so I intend to run everything under `graph` in the foreseeable future.