@jbduncan 2024-04-06
@kortschak Here are the benchmark results.

It looks like `Coloring/queen7_7/DsaturExact-8`, `AStarUndirected/NSW_Undirected_10_2_5_2-8` and `Coloring/exam_scheduler_2/WelshPowell-8` have the biggest variance in performance, so I'll re-run those with `-count=30` just in case.

```
goos: darwin
goarch: arm64
pkg: gonum.org/v1/gonum/graph/coloring
                                                       │ graph/benchmarks_graph_1948_old.txt │  graph/benchmarks_graph_1948_new.txt  │
                                                       │               sec/op                │    sec/op      vs base                │
Coloring/empty/Dsatur-8                                                         3.155n ±  0%    3.134n ±  0%   -0.67% (p=0.000 n=10)
Coloring/empty/DsaturExact-8                                                    3.167n ±  1%    3.149n ±  0%   -0.57% (p=0.000 n=10)
Coloring/empty/RecursiveLargestFirst-8                                          3.163n ±  0%    3.151n ±  0%   -0.40% (p=0.001 n=10)
Coloring/empty/SanSegundo-8                                                     3.159n ±  1%    3.151n ±  0%   -0.24% (p=0.013 n=10)
Coloring/empty/WelshPowell-8                                                    3.163n ±  0%    3.155n ±  0%   -0.25% (p=0.000 n=10)
Coloring/singleton/Dsatur-8                                                     406.7n ±  0%    407.3n ±  0%   +0.16% (p=0.017 n=10)
Coloring/singleton/DsaturExact-8                                                943.6n ±  2%    940.9n ±  0%   -0.30% (p=0.000 n=10)
Coloring/singleton/RecursiveLargestFirst-8                                      293.9n ±  0%    293.1n ±  0%   -0.27% (p=0.003 n=10)
Coloring/singleton/SanSegundo-8                                                 407.1n ±  0%    406.9n ±  0%        ~ (p=0.341 n=10)
Coloring/singleton/WelshPowell-8                                                284.2n ±  0%    284.0n ±  0%   -0.09% (p=0.005 n=10)
Coloring/kite/Dsatur-8                                                          3.266µ ±  0%    3.260µ ±  0%   -0.17% (p=0.002 n=10)
Coloring/kite/DsaturExact-8                                                     17.66µ ±  0%    17.61µ ±  0%   -0.24% (p=0.002 n=10)
Coloring/kite/RecursiveLargestFirst-8                                           2.223µ ±  0%    2.216µ ±  0%   -0.31% (p=0.000 n=10)
Coloring/kite/SanSegundo-8                                                      3.974µ ±  0%    3.965µ ±  0%   -0.24% (p=0.000 n=10)
Coloring/kite/WelshPowell-8                                                     1.387µ ±  0%    1.383µ ±  0%   -0.25% (p=0.001 n=10)
Coloring/triangle+1/Dsatur-8                                                    2.068µ ±  1%    2.064µ ±  0%   -0.22% (p=0.000 n=10)
Coloring/triangle+1/DsaturExact-8                                               9.957µ ±  0%    9.936µ ±  0%   -0.21% (p=0.008 n=10)
Coloring/triangle+1/RecursiveLargestFirst-8                                     1.438µ ±  1%    1.434µ ±  0%   -0.28% (p=0.000 n=10)
Coloring/triangle+1/SanSegundo-8                                                2.404µ ±  0%    2.400µ ±  0%        ~ (p=0.209 n=10)
Coloring/triangle+1/WelshPowell-8                                               966.4n ±  0%    961.6n ±  0%   -0.50% (p=0.000 n=10)
Coloring/bipartite_halves/Dsatur-8                                              7.945µ ±  0%    7.924µ ±  0%   -0.26% (p=0.015 n=10)
Coloring/bipartite_halves/DsaturExact-8                                         32.79µ ±  0%    32.64µ ±  0%   -0.45% (p=0.001 n=10)
Coloring/bipartite_halves/RecursiveLargestFirst-8                               3.969µ ±  0%    3.954µ ±  0%   -0.37% (p=0.000 n=10)
Coloring/bipartite_halves/SanSegundo-8                                          15.04µ ±  0%    14.97µ ±  0%   -0.46% (p=0.000 n=10)
Coloring/bipartite_halves/WelshPowell-8                                         2.381µ ±  0%    2.369µ ±  0%   -0.50% (p=0.000 n=10)
Coloring/bipartite_alternating/Dsatur-8                                         7.823µ ±  0%    7.797µ ±  0%   -0.33% (p=0.018 n=10)
Coloring/bipartite_alternating/DsaturExact-8                                    32.66µ ±  0%    32.53µ ±  0%   -0.40% (p=0.000 n=10)
Coloring/bipartite_alternating/RecursiveLargestFirst-8                          3.929µ ±  0%    3.921µ ±  0%   -0.19% (p=0.006 n=10)
Coloring/bipartite_alternating/SanSegundo-8                                     15.16µ ±  0%    15.08µ ±  0%   -0.54% (p=0.001 n=10)
Coloring/bipartite_alternating/WelshPowell-8                                    2.471µ ±  0%    2.461µ ±  0%   -0.40% (p=0.009 n=10)
Coloring/3/4_bipartite/Dsatur-8                                                 6.375µ ±  1%    6.342µ ±  0%   -0.52% (p=0.000 n=10)
Coloring/3/4_bipartite/DsaturExact-8                                            30.05µ ±  0%    29.95µ ±  0%   -0.35% (p=0.000 n=10)
Coloring/3/4_bipartite/RecursiveLargestFirst-8                                  3.510µ ±  0%    3.502µ ±  0%   -0.24% (p=0.000 n=10)
Coloring/3/4_bipartite/SanSegundo-8                                             13.23µ ±  0%    13.18µ ±  0%   -0.38% (p=0.000 n=10)
Coloring/3/4_bipartite/WelshPowell-8                                            2.225µ ±  0%    2.218µ ±  0%        ~ (p=0.078 n=10)
Coloring/cubical/Dsatur-8                                                       7.974µ ±  0%    7.931µ ±  0%   -0.54% (p=0.000 n=10)
Coloring/cubical/DsaturExact-8                                                  33.03µ ±  0%    32.94µ ±  0%   -0.27% (p=0.001 n=10)
Coloring/cubical/RecursiveLargestFirst-8                                        4.044µ ±  0%    4.027µ ±  0%   -0.40% (p=0.000 n=10)
Coloring/cubical/SanSegundo-8                                                   15.24µ ±  0%    15.18µ ±  1%   -0.44% (p=0.014 n=10)
Coloring/cubical/WelshPowell-8                                                  2.497µ ±  0%    2.486µ ±  0%   -0.44% (p=0.000 n=10)
Coloring/HoG33106/Dsatur-8                                                      18.97µ ±  0%    18.89µ ±  0%   -0.42% (p=0.000 n=10)
Coloring/HoG33106/DsaturExact-8                                                 72.86µ ±  0%    72.63µ ±  0%   -0.31% (p=0.000 n=10)
Coloring/HoG33106/RecursiveLargestFirst-8                                       8.829µ ±  0%    8.797µ ±  0%   -0.36% (p=0.012 n=10)
Coloring/HoG33106/SanSegundo-8                                                  68.46µ ±  0%    68.08µ ±  0%   -0.54% (p=0.001 n=10)
Coloring/HoG33106/WelshPowell-8                                                 4.971µ ±  0%    4.955µ ±  0%   -0.33% (p=0.002 n=10)
Coloring/HoG41237/Dsatur-8                                                      215.6m ±  1%    214.4m ±  0%   -0.57% (p=0.005 n=10)
Coloring/HoG41237/DsaturExact-8                                                 237.3m ±  0%    235.7m ±  0%   -0.67% (p=0.000 n=10)
Coloring/HoG41237/RecursiveLargestFirst-8                                       8.832m ±  1%    8.792m ±  1%   -0.46% (p=0.000 n=10)
Coloring/HoG41237/SanSegundo-8                                                  505.8m ±  1%    503.2m ±  0%   -0.51% (p=0.029 n=10)
Coloring/HoG41237/WelshPowell-8                                                 3.667m ±  0%    3.648m ±  1%        ~ (p=0.052 n=10)
Coloring/exams/Dsatur-8                                                         4.632µ ±  0%    4.617µ ±  0%   -0.33% (p=0.000 n=10)
Coloring/exams/DsaturExact-8                                                    26.41µ ±  0%    26.38µ ±  0%        ~ (p=0.118 n=10)
Coloring/exams/RecursiveLargestFirst-8                                          3.050µ ±  0%    3.037µ ±  0%   -0.41% (p=0.000 n=10)
Coloring/exams/SanSegundo-8                                                     7.402µ ±  0%    7.384µ ±  0%   -0.24% (p=0.008 n=10)
Coloring/exams/WelshPowell-8                                                    1.822µ ±  0%    1.818µ ±  0%   -0.22% (p=0.019 n=10)
Coloring/exam_scheduler_1/Dsatur-8                                              17.85µ ±  0%    17.81µ ±  0%   -0.27% (p=0.000 n=10)
Coloring/exam_scheduler_1/DsaturExact-8                                         98.88µ ±  0%    98.74µ ±  0%   -0.14% (p=0.043 n=10)
Coloring/exam_scheduler_1/RecursiveLargestFirst-8                               11.40µ ±  0%    11.34µ ±  0%   -0.51% (p=0.000 n=10)
Coloring/exam_scheduler_1/SanSegundo-8                                          27.39µ ±  0%    27.27µ ±  0%   -0.46% (p=0.004 n=10)
Coloring/exam_scheduler_1/WelshPowell-8                                         4.724µ ±  0%    4.710µ ±  0%   -0.30% (p=0.001 n=10)
Coloring/exam_scheduler_2/Dsatur-8                                              8.334µ ±  0%    8.300µ ±  0%   -0.40% (p=0.000 n=10)
Coloring/exam_scheduler_2/DsaturExact-8                                         47.88µ ±  0%    47.73µ ±  0%   -0.30% (p=0.000 n=10)
Coloring/exam_scheduler_2/RecursiveLargestFirst-8                               5.171µ ±  0%    5.174µ ±  0%        ~ (p=0.541 n=10)
Coloring/exam_scheduler_2/SanSegundo-8                                          12.74µ ±  0%    12.75µ ±  3%        ~ (p=0.493 n=10)
Coloring/exam_scheduler_2/WelshPowell-8                                         2.736µ ±  0%    2.927µ ±  3%   +6.96% (p=0.000 n=10)
Coloring/Brélaz/Dsatur-8                                                        12.04µ ±  0%    12.15µ ±  4%   +0.91% (p=0.000 n=10)
Coloring/Brélaz/DsaturExact-8                                                   59.31µ ±  0%    59.40µ ±  0%        ~ (p=0.165 n=10)
Coloring/Brélaz/RecursiveLargestFirst-8                                         7.580µ ±  0%    7.578µ ±  0%        ~ (p=0.868 n=10)
Coloring/Brélaz/SanSegundo-8                                                    18.32µ ±  0%    18.31µ ±  0%        ~ (p=0.493 n=10)
Coloring/Brélaz/WelshPowell-8                                                   3.708µ ±  0%    3.700µ ±  0%   -0.23% (p=0.000 n=10)
Coloring/Lima_Carmo/Dsatur-8                                                    7.533µ ±  0%    7.505µ ±  0%   -0.37% (p=0.000 n=10)
Coloring/Lima_Carmo/DsaturExact-8                                               32.62µ ±  0%    32.52µ ±  1%        ~ (p=0.101 n=10)
Coloring/Lima_Carmo/RecursiveLargestFirst-8                                     4.173µ ±  0%    4.161µ ±  0%   -0.29% (p=0.004 n=10)
Coloring/Lima_Carmo/SanSegundo-8                                                11.20µ ±  0%    11.16µ ±  0%   -0.29% (p=0.045 n=10)
Coloring/Lima_Carmo/WelshPowell-8                                               2.473µ ±  0%    2.465µ ±  0%   -0.32% (p=0.000 n=10)
Coloring/San_Segundo/Dsatur-8                                                   12.73µ ±  0%    12.71µ ±  1%        ~ (p=0.128 n=10)
Coloring/San_Segundo/DsaturExact-8                                              119.0µ ±  0%    118.8µ ±  0%   -0.15% (p=0.009 n=10)
Coloring/San_Segundo/RecursiveLargestFirst-8                                    9.210µ ±  0%    9.192µ ±  0%   -0.20% (p=0.006 n=10)
Coloring/San_Segundo/SanSegundo-8                                               22.39µ ±  0%    22.40µ ±  0%        ~ (p=0.594 n=10)
Coloring/San_Segundo/WelshPowell-8                                              3.602µ ±  2%    3.581µ ±  0%   -0.58% (p=0.000 n=10)
Coloring/tetrahedron/Dsatur-8                                                   2.378µ ±  0%    2.366µ ±  1%   -0.48% (p=0.021 n=10)
Coloring/tetrahedron/DsaturExact-8                                              13.58µ ±  0%    13.49µ ±  0%   -0.65% (p=0.000 n=10)
Coloring/tetrahedron/RecursiveLargestFirst-8                                    1.854µ ±  0%    1.847µ ±  0%   -0.40% (p=0.001 n=10)
Coloring/tetrahedron/SanSegundo-8                                               3.310µ ±  0%    3.294µ ±  0%   -0.48% (p=0.000 n=10)
Coloring/tetrahedron/WelshPowell-8                                              1.064µ ±  0%    1.057µ ±  0%   -0.66% (p=0.000 n=10)
Coloring/triangle/Dsatur-8                                                      1.477µ ±  0%    1.469µ ±  0%   -0.51% (p=0.001 n=10)
Coloring/triangle/DsaturExact-8                                                 7.236µ ±  0%    7.196µ ±  0%   -0.55% (p=0.000 n=10)
Coloring/triangle/RecursiveLargestFirst-8                                       1.152µ ±  0%    1.148µ ±  0%   -0.39% (p=0.001 n=10)
Coloring/triangle/SanSegundo-8                                                  1.818µ ±  0%    1.811µ ±  0%   -0.41% (p=0.001 n=10)
Coloring/triangle/WelshPowell-8                                                 768.6n ±  0%    765.2n ±  0%   -0.45% (p=0.001 n=10)
Coloring/square/Dsatur-8                                                        2.195µ ±  0%    2.187µ ±  1%        ~ (p=0.135 n=10)
Coloring/square/DsaturExact-8                                                   10.82µ ±  0%    10.78µ ±  0%   -0.32% (p=0.000 n=10)
Coloring/square/RecursiveLargestFirst-8                                         1.441µ ±  0%    1.435µ ±  0%   -0.42% (p=0.000 n=10)
Coloring/square/SanSegundo-8                                                    2.808µ ±  0%    2.797µ ±  0%   -0.39% (p=0.000 n=10)
Coloring/square/WelshPowell-8                                                   1.019µ ±  0%    1.016µ ±  0%   -0.34% (p=0.000 n=10)
Coloring/cycle-5/Dsatur-8                                                       3.220µ ±  0%    3.207µ ±  0%   -0.39% (p=0.004 n=10)
Coloring/cycle-5/DsaturExact-8                                                  19.15µ ±  0%    19.07µ ±  0%   -0.41% (p=0.000 n=10)
Coloring/cycle-5/RecursiveLargestFirst-8                                        2.117µ ±  0%    2.109µ ±  0%   -0.40% (p=0.000 n=10)
Coloring/cycle-5/SanSegundo-8                                                   4.143µ ±  0%    4.137µ ±  0%        ~ (p=0.101 n=10)
Coloring/cycle-5/WelshPowell-8                                                  1.360µ ±  0%    1.364µ ±  0%   +0.29% (p=0.011 n=10)
Coloring/cycle-6/Dsatur-8                                                       4.311µ ±  0%    4.329µ ±  0%   +0.41% (p=0.002 n=10)
Coloring/cycle-6/DsaturExact-8                                                  18.23µ ±  0%    18.33µ ±  0%   +0.54% (p=0.000 n=10)
Coloring/cycle-6/RecursiveLargestFirst-8                                        2.489µ ±  0%    2.489µ ±  1%        ~ (p=0.668 n=10)
Coloring/cycle-6/SanSegundo-8                                                   5.572µ ±  0%    5.571µ ±  0%        ~ (p=0.402 n=10)
Coloring/cycle-6/WelshPowell-8                                                  1.665µ ±  0%    1.661µ ±  1%        ~ (p=0.147 n=10)
Coloring/wheel-5/Dsatur-8                                                       3.443µ ±  0%    3.434µ ±  0%   -0.26% (p=0.001 n=10)
Coloring/wheel-5/DsaturExact-8                                                  22.79µ ±  0%    22.76µ ±  0%        ~ (p=0.109 n=10)
Coloring/wheel-5/RecursiveLargestFirst-8                                        2.386µ ±  0%    2.382µ ±  0%   -0.19% (p=0.001 n=10)
Coloring/wheel-5/SanSegundo-8                                                   5.165µ ±  0%    5.151µ ±  0%   -0.25% (p=0.002 n=10)
Coloring/wheel-5/WelshPowell-8                                                  1.395µ ±  0%    1.393µ ±  0%   -0.18% (p=0.039 n=10)
Coloring/wheel-6/Dsatur-8                                                       4.835µ ±  0%    4.822µ ±  0%   -0.26% (p=0.000 n=10)
Coloring/wheel-6/DsaturExact-8                                                  37.85µ ±  0%    37.74µ ±  0%   -0.30% (p=0.001 n=10)
Coloring/wheel-6/RecursiveLargestFirst-8                                        3.317µ ±  0%    3.310µ ±  0%   -0.23% (p=0.022 n=10)
Coloring/wheel-6/SanSegundo-8                                                   7.526µ ±  0%    7.505µ ±  0%   -0.27% (p=0.004 n=10)
Coloring/wheel-6/WelshPowell-8                                                  1.805µ ±  0%    1.802µ ±  0%   -0.17% (p=0.040 n=10)
Coloring/wheel-7/Dsatur-8                                                       6.316µ ±  0%    6.307µ ±  0%   -0.14% (p=0.024 n=10)
Coloring/wheel-7/DsaturExact-8                                                  39.24µ ±  0%    39.15µ ±  0%   -0.22% (p=0.005 n=10)
Coloring/wheel-7/RecursiveLargestFirst-8                                        3.949µ ±  0%    3.949µ ±  0%        ~ (p=1.000 n=10)
Coloring/wheel-7/SanSegundo-8                                                   10.16µ ±  0%    10.13µ ±  0%   -0.30% (p=0.011 n=10)
Coloring/wheel-7/WelshPowell-8                                                  2.179µ ±  0%    2.176µ ±  0%   -0.14% (p=0.000 n=10)
Coloring/sudoku_board/Dsatur-8                                                  21.63m ±  0%    21.58m ±  1%        ~ (p=0.190 n=10)
Coloring/sudoku_board/DsaturExact-8                                             86.28m ±  0%    85.99m ±  1%        ~ (p=0.218 n=10)
Coloring/sudoku_board/RecursiveLargestFirst-8                                   3.922m ±  0%    3.911m ±  0%   -0.29% (p=0.004 n=10)
Coloring/sudoku_board/SanSegundo-8                                              25.58m ±  0%    25.54m ±  0%        ~ (p=0.280 n=10)
Coloring/sudoku_board/WelshPowell-8                                             693.5µ ±  0%    692.9µ ±  0%        ~ (p=0.052 n=10)
Coloring/sudoku_problem/Dsatur-8                                                32.82m ±  0%    32.75m ±  0%   -0.22% (p=0.007 n=10)
Coloring/sudoku_problem/DsaturExact-8                                            6.062 ± 73%     6.569 ± 48%        ~ (p=0.739 n=10)
Coloring/sudoku_problem/RecursiveLargestFirst-8                                 5.860m ±  0%    5.833m ±  0%   -0.46% (p=0.001 n=10)
Coloring/sudoku_problem/SanSegundo-8                                            35.85m ±  0%    35.76m ±  0%   -0.26% (p=0.001 n=10)
Coloring/sudoku_problem/WelshPowell-8                                           974.1µ ±  0%    971.4µ ±  0%   -0.28% (p=0.003 n=10)
Coloring/cs_shc/Dsatur-8                                                        3.348µ ±  0%    3.346µ ±  0%        ~ (p=0.301 n=10)
Coloring/cs_shc/DsaturExact-8                                                   20.31µ ±  1%    20.28µ ±  0%   -0.15% (p=0.001 n=10)
Coloring/cs_shc/RecursiveLargestFirst-8                                         2.379µ ±  0%    2.378µ ±  0%        ~ (p=0.180 n=10)
Coloring/cs_shc/SanSegundo-8                                                    5.104µ ±  0%    5.099µ ±  0%        ~ (p=0.147 n=10)
Coloring/cs_shc/WelshPowell-8                                                   1.420µ ±  0%    1.417µ ±  0%        ~ (p=0.101 n=10)
Coloring/gis_hc/Dsatur-8                                                        3.973µ ±  0%    3.967µ ±  0%        ~ (p=0.078 n=10)
Coloring/gis_hc/DsaturExact-8                                                   16.74µ ±  1%    16.71µ ±  0%        ~ (p=0.256 n=10)
Coloring/gis_hc/RecursiveLargestFirst-8                                         2.312µ ±  0%    2.310µ ±  0%        ~ (p=0.075 n=10)
Coloring/gis_hc/SanSegundo-8                                                    7.445µ ±  0%    7.438µ ±  0%        ~ (p=0.137 n=10)
Coloring/gis_hc/WelshPowell-8                                                   1.648µ ±  0%    1.644µ ±  0%   -0.24% (p=0.002 n=10)
Coloring/gis_shc|rs_shc/Dsatur-8                                                2.072µ ±  0%    2.067µ ±  0%   -0.24% (p=0.000 n=10)
Coloring/gis_shc|rs_shc/DsaturExact-8                                           9.709µ ±  0%    9.685µ ±  0%   -0.24% (p=0.007 n=10)
Coloring/gis_shc|rs_shc/RecursiveLargestFirst-8                                 1.343µ ±  0%    1.339µ ±  0%   -0.30% (p=0.000 n=10)
Coloring/gis_shc|rs_shc/SanSegundo-8                                            2.699µ ±  0%    2.693µ ±  0%   -0.22% (p=0.031 n=10)
…